# Python __pycache__/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.0
pytelegrambotapi==4.18.0
gunicorn==21.2.0
pillow==10.2.0
requests==2.31.0